requirements.txt
asgiref==3.8.1
attrs==25.3.0
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
dj-database-url==3.0.0
Django==5.2.3
et_xmlfile==2.0.0
h11==0.16.0
idna==3.10
numpy==2.2.5
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.2.3
psycopg2-binary==2.9.10
pycparser==2.22
PyPDF2==3.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
selenium==4.34.2
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
sqlparse==0.5.3
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
